separate resource mutation from server code
