+++ pkg/webhooks/log.go
@@ -1,0 +1,5 @@
+package webhooks
+
+import "sigs.k8s.io/controller-runtime/pkg/log"
+
+var logger = log.Log.WithName("webhooks")
