+++ pkg/webhooks/resource/generation_test.go
@@ -1,4 +1,4 @@
-package webhooks
+package resource
 
 import (
 	"reflect"
