+++ pkg/webhooks/resource/validate_audit.go
@@ -1,4 +1,4 @@
-package webhooks
+package resource
 
 import (
 	"strings"
