+++ pkg/webhooks/resource/validation_test.go
@@ -1,4 +1,4 @@
-package webhooks
+package resource
 
 import (
 	"encoding/json"
